Update.
2000-11-28  Ulrich Drepper  <drepper@redhat.com>

	* time/strftime.c: Add const where necessary to avoid warnings.
	* time/strptime.c (strptime_internal): Add casts to avoid warnings.

+++ time/strptime.c
@@ -725,7 +725,7 @@ strptime_internal (rp, fmt, tm, decided, era_cnt)
 		    {
 		      num_eras = _NL_CURRENT_WORD (LC_TIME,
 						   _NL_TIME_ERA_NUM_ENTRIES);
-		      for (era_cnt = 0; era_cnt < num_eras;
+		      for (era_cnt = 0; era_cnt < (int) num_eras;
 			   ++era_cnt, rp = rp_backup)
 			{
 			  era = _nl_select_era_entry (era_cnt);
@@ -735,7 +735,7 @@ strptime_internal (rp, fmt, tm, decided, era_cnt)
 			      break;
 			    }
 			}
-		      if (era_cnt == num_eras)
+		      if (era_cnt == (int) num_eras)
 			{
 			  era_cnt = -1;
 			  if (*decided == loc)
@@ -764,14 +764,14 @@ strptime_internal (rp, fmt, tm, decided, era_cnt)
 		{
 		  num_eras = _NL_CURRENT_WORD (LC_TIME,
 					       _NL_TIME_ERA_NUM_ENTRIES);
-		  for (era_cnt = 0; era_cnt < num_eras;
+		  for (era_cnt = 0; era_cnt < (int) num_eras;
 		       ++era_cnt, rp = rp_backup)
 		    {
 		      era = _nl_select_era_entry (era_cnt);
 		      if (recursive (era->era_format))
 			break;
 		    }
-		  if (era_cnt == num_eras)
+		  if (era_cnt == (int) num_eras)
 		    {
 		      era_cnt = -1;
 		      if (*decided == loc)
